+++ bootique/src/main/java/com/nhl/bootique/config/jackson/InPlaceMapOverrider.java
@@ -15,16 +15,20 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 public class InPlaceMapOverrider implements Function<JsonNode, JsonNode> {
 
 	private Map<String, String> properties;
+	private boolean caseSensitive;
+	private char pathSeparator;
 
-	public InPlaceMapOverrider(Map<String, String> properties) {
+	public InPlaceMapOverrider(Map<String, String> properties, boolean caseSensitive, char pathSeparator) {
 		this.properties = properties;
+		this.caseSensitive = caseSensitive;
+		this.pathSeparator = pathSeparator;
 	}
 
 	@Override
 	public JsonNode apply(JsonNode t) {
 		properties.entrySet().forEach(e -> {
 
-			PathSegment target = new PathSegment(t, e.getKey()).lastPathComponent().get();
+			PathSegment target = lastPathComponent(t, e.getKey());
 			target.fillMissingParents();
 
 			if (!(target.getParentNode() instanceof ObjectNode)) {
@@ -36,5 +40,11 @@ public class InPlaceMapOverrider implements Function<JsonNode, JsonNode> {
 		});
 
 		return t;
+	}
+
+	protected PathSegment lastPathComponent(JsonNode t, String path) {
+		PathSegment root = caseSensitive ? new PathSegment(t, path, pathSeparator)
+				: new CiPathSegment(t, path, pathSeparator);
+		return root.lastPathComponent().get();
 	}
 }