+++ bootique/src/main/java/com/nhl/bootique/env/Environment.java
@@ -2,6 +2,10 @@ package com.nhl.bootique.env;
 
 import java.util.Map;
 
+/**
+ * Provides access to system properties and environment variables for the app.
+ * Allows to filter properties by prefix to separate Bootique-specific values.
+ */
 public interface Environment {
 
 	String getProperty(String name);
@@ -25,4 +29,34 @@ public interface Environment {
 	 * @return a map of all properties that start with "bq." prefix.
 	 */
 	Map<String, String> frameworkProperties();
+
+	/**
+	 * Returns a value of the environment variable with a given name.
+	 * 
+	 * @since 0.17
+	 * @param name
+	 *            environment variable name.
+	 * @return a value of the environment variable with a given name.
+	 */
+	String getVariable(String name);
+
+	/**
+	 * Returns a map of environment variables that start with a prefix. Prefix
+	 * is stripped from the returned names.
+	 * 
+	 * @since 0.17
+	 * @param prefix
+	 *            a prefix to qualify variables with.
+	 * @return a map of environment variables that start with a prefix.
+	 */
+	Map<String, String> variables(String prefix);
+
+	/**
+	 * Returns a map of all variables that start with "BQ_" prefix. Prefix is
+	 * stripped from the returned names.
+	 * 
+	 * @since 0.17
+	 * @return a map of all variables that start with "BQ_" prefix.
+	 */
+	Map<String, String> frameworkVariables();
 }