+++ bootique/src/test/java/com/nhl/bootique/config/jackson/PathSegmentTest.java
@@ -31,7 +31,7 @@ public class PathSegmentTest {
 	@Test
 	public void testLastPathComponent_Root() {
 		JsonNode node = readYaml("a: b\nc: d");
-		Optional<PathSegment> last = new PathSegment(node, "").lastPathComponent();
+		Optional<PathSegment> last = new PathSegment(node, "", '.').lastPathComponent();
 
 		assertNotNull(last);
 		assertNotNull(last.get());
@@ -42,7 +42,7 @@ public class PathSegmentTest {
 	@Test
 	public void testLastPathComponent() {
 		JsonNode node = readYaml("a: b\nc: d");
-		Optional<PathSegment> last = new PathSegment(node, "a").lastPathComponent();
+		Optional<PathSegment> last = new PathSegment(node, "a", '.').lastPathComponent();
 
 		assertNotNull(last);
 		assertNotNull(last.get());
@@ -52,10 +52,9 @@ public class PathSegmentTest {
 	@Test
 	public void testLastPathComponent_Nested() {
 		JsonNode node = readYaml("a: b\nc:\n  d: e");
-		Optional<PathSegment> last = new PathSegment(node, "c.d").lastPathComponent();
+		Optional<PathSegment> last = new PathSegment(node, "c.d", '.').lastPathComponent();
 
 		assertNotNull(last);
-		assertNotNull(last.get());
 		assertEquals("e", last.get().getNode().asText());
 	}
 }