+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #62 Bootique.main() - a generic main method
 * #63 Must canonicalize file:/ URLs.
+* #65 Allow property override from environment variables
 * #66 Support for multiple ordered configs
 * #67 Expand Bootique.app(..) method